- Add a "FogType" graphical parameter. Offering : SkyFog, NormalFog, NoFog (= PopUP Chunk).
Been a long time I wanted to do it, mainly because my computer here at work is really
slow, and the currently working Sky fading effect with the landscape is killing the perf.

Normal Fog is the fastest mode (28% less heavy on the GPU than SkyFog)
NoFog behind the old "Chunk pop up" mode (25% less heavy but displaying more landscape)

Still need to configure the parameter to made it possible to do a change without restarting the
game.